add auth reopsitory mock tests
