+++ src/repositories/auth_repository.rs
@@ -116,111 +116,3 @@ impl PostgresDB {
             })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::{mem, ops::DerefMut, sync::Arc};
-
-    use anyhow::anyhow;
-    use tokio::sync::Mutex;
-
-    use crate::domain::{
-        model::{
-            auth::AuthorizationError,
-            login_user::{LoginUserError, LoginUserRequest},
-            register_user::{HashedUserPassword, RegisterUserError, RegisterUserRequest},
-            user::{FilteredUser, User},
-            user_email::UserEmail,
-            user_id::UserId,
-            user_password::UserPassword,
-        },
-        repositories::auth_repository::AuthRepository,
-    };
-
-    struct MockAuthRepository {
-        /// It would be great for result to just take a Result instead of the below, unfortunately
-        /// it needs to conform to `Clone` but RegisterUserError` has an `Unknown` variant that
-        /// might wrap errors that are not Clone.
-        register_result: Arc<Mutex<Result<FilteredUser, RegisterUserError>>>,
-        auth_result: Arc<Mutex<Result<User, AuthorizationError>>>,
-        login_result: Arc<Mutex<Result<User, LoginUserError>>>,
-    }
-
-    impl AuthRepository for MockAuthRepository {
-        async fn register(
-            &self,
-            _request: &RegisterUserRequest,
-        ) -> Result<FilteredUser, RegisterUserError> {
-            let mut guard = self.register_result.lock().await;
-            let mut result = Err(RegisterUserError::Unknown(anyhow!("substitute error")));
-            mem::swap(guard.deref_mut(), &mut result);
-            result
-        }
-
-        async fn auth(&self, _request: &UserId) -> Result<User, AuthorizationError> {
-            let mut guard = self.auth_result.lock().await;
-            let mut result = Err(AuthorizationError::Unknown(anyhow!("substitute error")));
-            mem::swap(guard.deref_mut(), &mut result);
-            result
-        }
-
-        async fn login(&self, _request: &LoginUserRequest) -> Result<User, LoginUserError> {
-            let mut guard = self.login_result.lock().await;
-            let mut result = Err(LoginUserError::Unknown(anyhow!("substitute error")));
-            mem::swap(guard.deref_mut(), &mut result);
-            result
-        }
-    }
-
-    // TODO: - Remove sanity check tests
-    #[tokio::test]
-    async fn test_register_success() {
-        let email = "adrian@email.com";
-        let password = "password";
-        let user = User::new(email, password);
-        let filtered_user = FilteredUser::from(&user);
-        let register_result = Arc::new(Mutex::new(Ok(filtered_user)));
-        let auth_result = Arc::new(Mutex::new(Ok(user.clone())));
-        let login_result = Arc::new(Mutex::new(Ok(user)));
-
-        let mock_repo = MockAuthRepository {
-            register_result,
-            auth_result,
-            login_result,
-        };
-
-        let register = mock_repo
-            .register(&RegisterUserRequest::new(
-                UserEmail::new(email).unwrap(),
-                HashedUserPassword::new(UserPassword::new(password).unwrap()).unwrap(),
-            ))
-            .await;
-
-        assert_eq!(email.to_string(), register.unwrap().email);
-    }
-
-    #[tokio::test]
-    async fn test_auth_success() {
-        let email = "adrian@email.com";
-        let password = "password";
-        let mut user = User::new(email, password);
-        let uuid = uuid::Uuid::new_v4();
-        user.id = uuid;
-        let filtered_user = FilteredUser::from(&user);
-        let register_result = Arc::new(Mutex::new(Ok(filtered_user)));
-        let auth_result = Arc::new(Mutex::new(Ok(user.clone())));
-        let login_result = Arc::new(Mutex::new(Ok(user)));
-
-        let user_id = UserId::new(uuid);
-
-        let mock_repo = MockAuthRepository {
-            register_result,
-            auth_result,
-            login_result,
-        };
-
-        let register = mock_repo.auth(&user_id).await;
-
-        assert_eq!(user_id.get(), &register.unwrap().id);
-    }
-}
