+++ src/repositories/mod.rs
@@ -1,2 +1,3 @@
 pub mod auth_repository;
 pub mod cache_repository;
+pub mod test_helpers;
